requirements.txt
# FastAPI & ASGI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlmodel==0.0.22
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2
greenlet==3.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
httpx==0.27.2

# Celery & Redis
celery[redis]==5.4.0
redis==5.1.0

# Cloud Storage
google-cloud-storage==2.18.2

# AI/ML APIs
openai==1.51.0
anthropic==0.34.2

# Utilities
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.2.0
tenacity==9.0.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
